Split tests (too long for CI)
