+++ tests/ui/search_is_some_fixable_none.rs
@@ -6,45 +6,6 @@ fn main() {
     let v = vec![3, 2, 1, 0, -1, -2, -3];
     let y = &&42;
 
-    // Check `find().is_some()`, single-line case.
-    let _ = v.iter().find(|&x| *x < 0).is_some();
-    let _ = (0..1).find(|x| **y == *x).is_some(); // one dereference less
-    let _ = (0..1).find(|x| *x == 0).is_some();
-    let _ = v.iter().find(|x| **x == 0).is_some();
-    let _ = (4..5).find(|x| *x == 1 || *x == 3 || *x == 5).is_some();
-    let _ = (1..3).find(|x| [1, 2, 3].contains(x)).is_some();
-    let _ = (1..3).find(|x| *x == 0 || [1, 2, 3].contains(x)).is_some();
-    let _ = (1..3).find(|x| [1, 2, 3].contains(x) || *x == 0).is_some();
-    let _ = (1..3)
-        .find(|x| [1, 2, 3].contains(x) || *x == 0 || [4, 5, 6].contains(x) || *x == -1)
-        .is_some();
-
-    // Check `position().is_some()`, single-line case.
-    let _ = v.iter().position(|&x| x < 0).is_some();
-
-    // Check `rposition().is_some()`, single-line case.
-    let _ = v.iter().rposition(|&x| x < 0).is_some();
-
-    let s1 = String::from("hello world");
-    let s2 = String::from("world");
-    // caller of `find()` is a `&`static str`
-    let _ = "hello world".find("world").is_some();
-    let _ = "hello world".find(&s2).is_some();
-    let _ = "hello world".find(&s2[2..]).is_some();
-    // caller of `find()` is a `String`
-    let _ = s1.find("world").is_some();
-    let _ = s1.find(&s2).is_some();
-    let _ = s1.find(&s2[2..]).is_some();
-    // caller of `find()` is slice of `String`
-    let _ = s1[2..].find("world").is_some();
-    let _ = s1[2..].find(&s2).is_some();
-    let _ = s1[2..].find(&s2[2..]).is_some();
-}
-
-fn is_none() {
-    let v = vec![3, 2, 1, 0, -1, -2, -3];
-    let y = &&42;
-
     // Check `find().is_none()`, single-line case.
     let _ = v.iter().find(|&x| *x < 0).is_none();
     let _ = (0..1).find(|x| **y == *x).is_none(); // one dereference less
